+++ cedar-policy-core/src/entities.rs
@@ -352,6 +352,7 @@ mod json_parsing_tests {
         )
     }
 
+    #[cfg(feature = "ipaddr")]
     /// this one uses `__expr`, `__entity`, and `__extn` escapes, in various positions
     #[test]
     fn more_escapes() {
@@ -964,6 +965,7 @@ mod schema_based_parsing_tests {
         }
     }
 
+    #[cfg(feature = "ipaddr")]
     /// JSON that should parse differently with and without the above schema
     #[test]
     fn with_and_without_schema() {
@@ -1158,6 +1160,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// simple type mismatch with expected type
     #[test]
     fn type_mismatch_string_long() {
@@ -1203,6 +1206,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// another simple type mismatch with expected type
     #[test]
     fn type_mismatch_entity_record() {
@@ -1249,6 +1253,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// type mismatch where we expect a set and get just a single element
     #[test]
     fn type_mismatch_set_element() {
@@ -1291,6 +1296,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// type mismatch where we just get the wrong entity type
     #[test]
     fn type_mismatch_entity_types() {
@@ -1336,6 +1342,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// type mismatch where we're expecting an extension type and get a
     /// different extension type
     #[test]
@@ -1382,6 +1389,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     #[test]
     fn missing_record_attr() {
         // missing a record attribute entirely
@@ -1426,6 +1434,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// record attribute has the wrong type
     #[test]
     fn type_mismatch_in_record_attr() {
@@ -1502,6 +1511,7 @@ mod schema_based_parsing_tests {
             .expect("this version with explicit __entity and __extn escapes should also pass");
     }
 
+    #[cfg(feature = "ipaddr")]
     /// unexpected record attribute
     #[test]
     fn unexpected_record_attr() {
@@ -1592,6 +1602,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// unexpected entity attribute
     #[test]
     fn unexpected_entity_attr() {
@@ -1640,6 +1651,7 @@ mod schema_based_parsing_tests {
         );
     }
 
+    #[cfg(feature = "ipaddr")]
     /// Test that involves parents of wrong types
     #[test]
     fn parents_wrong_type() {