+++ cedar-policy/src/api.rs
@@ -144,6 +144,8 @@ impl std::fmt::Display for Entity {
 #[derive(Debug, Clone, Default, PartialEq, Eq, RefCast)]
 pub struct Entities(pub(crate) entities::Entities);
 
+pub use entities::EntitiesError;
+
 impl Entities {
     /// Create a fresh `Entities` with no entities
     pub fn empty() -> Self {
@@ -315,20 +317,8 @@ impl Authorizer {
             .0
             .is_authorized_core(&query.0, &policy_set.ast, &entities.0);
         match response {
-            authorizer::ResponseKind::FullyEvaluated(a) => PartialResponse::Concrete(Response {
-                decision: a.decision,
-                diagnostics: Diagnostics {
-                    reason: a.diagnostics.reason.into_iter().map(PolicyId).collect(),
-                    errors: a.diagnostics.errors.into_iter().collect(),
-                },
-            }),
-            authorizer::ResponseKind::Partial(p) => PartialResponse::Residual(ResidualResponse {
-                residuals: PolicySet::from_ast(p.residuals),
-                diagnostics: Diagnostics {
-                    reason: p.diagnostics.reason.into_iter().map(PolicyId).collect(),
-                    errors: p.diagnostics.errors.into_iter().collect(),
-                },
-            }),
+            authorizer::ResponseKind::FullyEvaluated(a) => PartialResponse::Concrete(a.into()),
+            authorizer::ResponseKind::Partial(p) => PartialResponse::Residual(p.into()),
         }
     }
 }
@@ -373,6 +363,15 @@ pub struct Diagnostics {
     errors: HashSet<String>,
 }
 
+impl From<authorizer::Diagnostics> for Diagnostics {
+    fn from(diagnostics: authorizer::Diagnostics) -> Self {
+        Self {
+            reason: diagnostics.reason.into_iter().map(PolicyId).collect(),
+            errors: diagnostics.errors.iter().map(ToString::to_string).collect(),
+        }
+    }
+}
+
 impl Diagnostics {
     /// Get the policies that contributed to the decision
     pub fn reason(&self) -> impl Iterator<Item = &PolicyId> {
@@ -412,10 +411,7 @@ impl From<authorizer::Response> for Response {
     fn from(a: authorizer::Response) -> Self {
         Self {
             decision: a.decision,
-            diagnostics: Diagnostics {
-                reason: a.diagnostics.reason.into_iter().map(PolicyId).collect(),
-                errors: a.diagnostics.errors.into_iter().collect(),
-            },
+            diagnostics: a.diagnostics.into(),
         }
     }
 }
@@ -438,6 +434,16 @@ impl ResidualResponse {
     /// Get the authorization diagnostics
     pub fn diagnostics(&self) -> &Diagnostics {
         &self.diagnostics
+    }
+}
+
+#[cfg(feature = "partial-eval")]
+impl From<authorizer::PartialResponse> for ResidualResponse {
+    fn from(p: authorizer::PartialResponse) -> Self {
+        Self {
+            residuals: PolicySet::from_ast(p.residuals),
+            diagnostics: p.diagnostics.into(),
+        }
     }
 }
 
@@ -1116,17 +1122,17 @@ impl FromStr for PolicySet {
     ///
     /// See [`Policy`] for more.
     fn from_str(policies: &str) -> Result<Self, Self::Err> {
-        let (ests, pset) = parser::parse_policyset_to_ests_and_pset(policies)?;
+        let (texts, pset) = parser::parse_policyset_and_also_return_policy_text(policies)?;
         let policies = pset.policies().map(|p|
             (
                 PolicyId(p.id().clone()),
-                Policy { est: ests.get(p.id()).expect("internal invariant violation: policy id exists in asts but not ests").clone(), ast: p.clone() }
+                Policy { lossless: LosslessPolicy::policy_or_template_text(*texts.get(p.id()).expect("internal invariant violation: policy id exists in asts but not texts")), ast: p.clone() }
             )
         ).collect();
         let templates = pset.templates().map(|t|
             (
                 PolicyId(t.id().clone()),
-                Template { est: ests.get(t.id()).expect("internal invariant violation: template id exists in asts but not ests").clone(), ast: t.clone() }
+                Template { lossless: LosslessPolicy::policy_or_template_text(*texts.get(t.id()).expect("internal invariant violation: template id exists in asts but not ests")), ast: t.clone() }
             )
         ).collect();
         Ok(Self {
@@ -1182,7 +1188,7 @@ impl PolicySet {
 
     /// Iterate over all the `Policy`s in the `PolicySet`.
     ///
-    /// This will include policies that result from instantiating a template, and inline policies.
+    /// This will include both static and template-linked policies.
     pub fn policies(&self) -> impl Iterator<Item = &Policy> {
         self.policies.values()
     }
@@ -1232,9 +1238,6 @@ impl PolicySet {
     /// Failure can happen for three reasons
     ///   1) The map passed in `vals` may not match the slots in the template
     ///   2) The `new_id` may conflict w/ a policy that already exists in the set
-    ///   3) `template_id` does not correspond to a template. Either the id is
-    ///   not in the policy set, or it is in the policy set but is either a
-    ///   linked or static policy rather than a template
     #[allow(clippy::needless_pass_by_value)]
     pub fn link(
         &mut self,
@@ -1242,21 +1245,22 @@ impl PolicySet {
         new_id: PolicyId,
         vals: HashMap<SlotId, EntityUid>,
     ) -> Result<(), PolicySetError> {
-        let unwrapped: HashMap<ast::SlotId, ast::EntityUID> = vals
+        let unwrapped_vals: HashMap<ast::SlotId, ast::EntityUID> = vals
             .into_iter()
             .map(|(key, value)| (key.into(), value.0))
             .collect();
-        let est_vals = unwrapped.iter().map(|(k, v)| (*k, v.into())).collect();
         self.ast
-            .link(template_id.0.clone(), new_id.0.clone(), unwrapped)
+            .link(
+                template_id.0.clone(),
+                new_id.0.clone(),
+                unwrapped_vals.clone(),
+            )
             .map_err(PolicySetError::LinkingError)?;
         let linked_ast = self
             .ast
             .get(&new_id.0)
-            .expect("instantiate() didn't fail above, so this shouldn't fail")
-            .clone();
-        //TODO Rename this
-        let lined_est = self
+            .expect("ast.link() didn't fail above, so this shouldn't fail");
+        let linked_lossless = self
             .templates
             .get(&template_id)
             // We know `template_id` exists in the policy set as either a
@@ -1265,9 +1269,9 @@ impl PolicySet {
             // be that the id corresponds to a static policy. This function
             // should only be used to link templates, so this is an error.
             .ok_or(PolicySetError::ExpectedTemplate)?
+            .lossless
             .clone()
-            .est
-            .link(&est_vals)
+            .link(unwrapped_vals.iter().map(|(k, v)| (*k, v)))
             // The only error case for `est.link()` is a template with
             // slots which are not filled by the provided values. `ast.link()`
             // will have already errored if there are any unfilled slots in the
@@ -1276,8 +1280,8 @@ impl PolicySet {
         self.policies.insert(
             new_id,
             Policy {
-                ast: linked_ast,
-                est: lined_est,
+                ast: linked_ast.clone(),
+                lossless: linked_lossless,
             },
         );
         Ok(())
@@ -1318,15 +1322,18 @@ pub struct Template {
     /// AST representation of the template, used for most operations.
     /// In particular, the `ast` contains the authoritative `PolicyId` for the template.
     ast: ast::Template,
-    /// EST representation of the template, used just for `to_json()`.
+
+    /// Some "lossless" representation of the template, whichever is most
+    /// convenient to provide (and can be provided with the least overhead).
+    /// This is used just for `to_json()`.
     /// We can't just derive this on-demand from `ast`, because the AST is lossy:
     /// we can't reconstruct an accurate CST/EST/policy-text from the AST, but
     /// we can from the EST (modulo whitespace and a few other things like the
     /// order of annotations).
     ///
-    /// This is an est::Policy because the EST doesn't distinguish between
-    /// inline policies and templates.
-    est: est::Policy,
+    /// This is a `LosslessPolicy` (rather than something like `LosslessTemplate`)
+    /// because the EST doesn't distinguish between static policies and templates.
+    lossless: LosslessPolicy,
 }
 
 impl PartialEq for Template {
@@ -1341,9 +1348,13 @@ impl Template {
     /// Attempt to parse a `Template` from source.
     /// If `id` is Some, then the resulting template will have that `id`.
     /// If the `id` is None, the parser will use the default "policy0".
+    /// The behavior around None may change in the future.
     pub fn parse(id: Option<String>, src: impl AsRef<str>) -> Result<Self, ParseErrors> {
-        let (est, ast) = parser::parse_policy_template_to_est_and_ast(id, src.as_ref())?;
-        Ok(Self { ast, est })
+        let ast = parser::parse_policy_template(id, src.as_ref()).map_err(ParseErrors)?;
+        Ok(Self {
+            ast,
+            lossless: LosslessPolicy::policy_or_template_text(src.as_ref()),
+        })
     }
 
     /// Get the `PolicyId` of this `Template`
@@ -1356,7 +1367,7 @@ impl Template {
     pub fn new_id(&self, id: PolicyId) -> Self {
         Self {
             ast: self.ast.new_id(id.0),
-            est: self.est.clone(),
+            lossless: self.lossless.clone(), // Lossless representation doesn't include the `PolicyId`
         }
     }
 
@@ -1449,14 +1460,16 @@ impl Template {
             serde_json::from_value(json).map_err(JsonDeserializationError::Serde)?;
         Ok(Self {
             ast: est.clone().try_into_ast_template(id.map(|id| id.0))?,
-            est,
+            lossless: LosslessPolicy::Est(est),
         })
     }
 
     /// Get the JSON representation of this `Template`.
     #[allow(dead_code)] // planned to be a public method in the future
     fn to_json(&self) -> Result<serde_json::Value, impl std::error::Error> {
-        serde_json::to_value(&self.est)
+        let est = self.lossless.est()?;
+        let json = serde_json::to_value(est)?;
+        Ok::<_, PolicyToJsonError>(json)
     }
 
     /// Create a `Template` from its AST representation only. The EST will
@@ -1466,8 +1479,11 @@ impl Template {
     /// not the original policy syntax.
     #[cfg_attr(not(feature = "partial-eval"), allow(unused))]
     fn from_ast(ast: ast::Template) -> Self {
-        let est = ast.clone().into();
-        Self { ast, est }
+        let text = ast.to_string(); // assume that pretty-printing is faster than `est::Policy::from(ast.clone())`; is that true?
+        Self {
+            ast,
+            lossless: LosslessPolicy::policy_or_template_text(text),
+        }
     }
 }
 
@@ -1584,12 +1600,14 @@ pub struct Policy {
     /// AST representation of the policy, used for most operations.
     /// In particular, the `ast` contains the authoritative `PolicyId` for the policy.
     ast: ast::Policy,
-    /// EST representation of the policy, used just for `to_json()`.
+    /// Some "lossless" representation of the policy, whichever is most
+    /// convenient to provide (and can be provided with the least overhead).
+    /// This is used just for `to_json()`.
     /// We can't just derive this on-demand from `ast`, because the AST is lossy:
     /// we can't reconstruct an accurate CST/EST/policy-text from the AST, but
     /// we can from the EST (modulo whitespace and a few other things like the
     /// order of annotations).
-    est: est::Policy,
+    lossless: LosslessPolicy,
 }
 
 impl PartialEq for Policy {
@@ -1601,8 +1619,8 @@ impl PartialEq for Policy {
 impl Eq for Policy {}
 
 impl Policy {
-    /// Get the `PolicyId` of the `Template` this is linked to of.
-    /// If this is an static policy, this will return `None`.
+    /// Get the `PolicyId` of the `Template` this is linked to.
+    /// If this is a static policy, this will return `None`.
     pub fn template_id(&self) -> Option<&PolicyId> {
         if self.is_static() {
             None
@@ -1640,7 +1658,7 @@ impl Policy {
     pub fn new_id(&self, id: PolicyId) -> Self {
         Self {
             ast: self.ast.new_id(id.0),
-            est: self.est.clone(),
+            lossless: self.lossless.clone(), // Lossless representation doesn't include the `PolicyId`
         }
     }
 
@@ -1709,9 +1727,12 @@ impl Policy {
     /// If `id` is None, then "policy0" will be used.
     /// The behavior around None may change in the future.
     pub fn parse(id: Option<String>, policy_src: impl AsRef<str>) -> Result<Self, ParseErrors> {
-        let (est, inline_ast) = parser::parse_policy_to_est_and_ast(id, policy_src.as_ref())?;
+        let inline_ast = parser::parse_policy(id, policy_src.as_ref()).map_err(ParseErrors)?;
         let (_, ast) = ast::Template::link_static_policy(inline_ast);
-        Ok(Self { ast, est })
+        Ok(Self {
+            ast,
+            lossless: LosslessPolicy::policy_or_template_text(policy_src.as_ref()),
+        })
     }
 
     /// Create a `Policy` from its JSON representation.
@@ -1726,24 +1747,29 @@ impl Policy {
             serde_json::from_value(json).map_err(JsonDeserializationError::Serde)?;
         Ok(Self {
             ast: est.clone().try_into_ast_policy(id.map(|id| id.0))?,
-            est,
+            lossless: LosslessPolicy::Est(est),
         })
     }
 
     /// Get the JSON representation of this `Policy`.
     pub fn to_json(&self) -> Result<serde_json::Value, impl std::error::Error> {
-        serde_json::to_value(&self.est)
+        let est = self.lossless.est()?;
+        let json = serde_json::to_value(est)?;
+        Ok::<_, PolicyToJsonError>(json)
     }
 
-    /// Create a `Policy` from its AST representation only. The EST will
-    /// reflect the AST structure. When possible, don't use this method and
-    /// create the EST from the policy text or CST instead, as the conversion
-    /// to AST is lossy. ESTs generated by this method will reflect the AST and
-    /// not the original policy syntax.
+    /// Create a `Policy` from its AST representation only. The `LosslessPolicy`
+    /// will reflect the AST structure. When possible, don't use this method and
+    /// create the `Policy` from the policy text, CST, or EST instead, as the
+    /// conversion to AST is lossy. ESTs for policies generated by this method
+    /// will reflect the AST and not the original policy syntax.
     #[cfg_attr(not(feature = "partial-eval"), allow(unused))]
     fn from_ast(ast: ast::Policy) -> Self {
-        let est = ast.clone().into();
-        Self { ast, est }
+        let text = ast.to_string(); // assume that pretty-printing is faster than `est::Policy::from(ast.clone())`; is that true?
+        Self {
+            ast,
+            lossless: LosslessPolicy::policy_or_template_text(text),
+        }
     }
 }
 
@@ -1765,6 +1791,85 @@ impl FromStr for Policy {
     fn from_str(policy: &str) -> Result<Self, Self::Err> {
         Self::parse(None, policy)
     }
+}
+
+/// See comments on `Policy` and `Template`.
+///
+/// This structure can be used for static policies, linked policies, and templates.
+#[derive(Debug, Clone)]
+enum LosslessPolicy {
+    /// EST representation
+    Est(est::Policy),
+    /// Text representation
+    Text {
+        /// actual policy text, of the policy or template
+        text: String,
+        /// For linked policies, map of slot to UID. Only linked policies have
+        /// this; static policies and (unlinked) templates have an empty map
+        /// here
+        slots: HashMap<ast::SlotId, ast::EntityUID>,
+    },
+}
+
+impl LosslessPolicy {
+    /// Create a new `LosslessPolicy` from the text of a policy or template.
+    fn policy_or_template_text(text: impl Into<String>) -> Self {
+        Self::Text {
+            text: text.into(),
+            slots: HashMap::new(),
+        }
+    }
+
+    /// Get the EST representation of this static policy, linked policy, or template
+    fn est(&self) -> Result<est::Policy, PolicyToJsonError> {
+        match self {
+            Self::Est(est) => Ok(est.clone()),
+            Self::Text { text, slots } => {
+                let est = parser::parse_policy_or_template_to_est(text)?;
+                if slots.is_empty() {
+                    Ok(est)
+                } else {
+                    let unwrapped_vals = slots.iter().map(|(k, v)| (*k, v.into())).collect();
+                    Ok(est.link(&unwrapped_vals)?)
+                }
+            }
+        }
+    }
+
+    fn link<'a>(
+        self,
+        vals: impl IntoIterator<Item = (ast::SlotId, &'a ast::EntityUID)>,
+    ) -> Result<Self, est::InstantiationError> {
+        match self {
+            Self::Est(est) => {
+                let unwrapped_est_vals: HashMap<ast::SlotId, entities::EntityUidJSON> =
+                    vals.into_iter().map(|(k, v)| (k, v.into())).collect();
+                Ok(Self::Est(est.link(&unwrapped_est_vals)?))
+            }
+            Self::Text { text, slots } => {
+                debug_assert!(
+                    slots.is_empty(),
+                    "shouldn't call link() on an already-linked policy"
+                );
+                let slots = vals.into_iter().map(|(k, v)| (k, v.clone())).collect();
+                Ok(Self::Text { text, slots })
+            }
+        }
+    }
+}
+
+/// Errors that can happen when getting the JSON representation of a policy
+#[derive(Debug, Error)]
+pub enum PolicyToJsonError {
+    /// Parse error in the policy text
+    #[error(transparent)]
+    Parse(#[from] ParseErrors),
+    /// For linked policies, error linking the JSON representation
+    #[error(transparent)]
+    Link(#[from] est::InstantiationError),
+    /// Error in the JSON serialization
+    #[error(transparent)]
+    Serde(#[from] serde_json::Error),
 }
 
 /// Expressions to be evaluated
@@ -2783,7 +2888,7 @@ mod schema_tests {
     /// A minimal test that a valid Schema parses
     #[test]
     fn valid_schema() {
-        let _ = Schema::from_json_value(json!(
+        Schema::from_json_value(json!(
         { "": {
             "entityTypes": {
                 "Photo": {
@@ -3356,7 +3461,8 @@ mod schema_based_parsing_tests {
                 }
             ]
         );
-        let _ = Entities::from_json_value(entitiesjson, Some(&schema))
+
+        Entities::from_json_value(entitiesjson, Some(&schema))
             .expect("this version with explicit __entity and __extn escapes should also pass");
     }
 
@@ -3804,6 +3910,7 @@ mod schema_based_parsing_tests {
             }
         }}))
         .unwrap();
+
         let schema = Schema::from_schema_fragments([fragment]).unwrap();
         let action_entities = schema.action_entities().unwrap();
 
@@ -3812,6 +3919,7 @@ mod schema_based_parsing_tests {
         let c_euid = EntityUid::from_strs("Action", "C");
         let d_euid = EntityUid::from_strs("Action", "D");
         let e_euid = EntityUid::from_strs("Action", "E");
+
         assert_eq!(
             action_entities,
             Entities::from_entities([
@@ -3832,12 +3940,12 @@ mod schema_based_parsing_tests {
                     HashSet::from([a_euid.clone(), b_euid.clone(), c_euid.clone()])
                 ),
                 Entity::new(
-                    e_euid.clone(),
+                    e_euid,
                     HashMap::new(),
                     HashSet::from([a_euid, b_euid, c_euid, d_euid])
                 ),
             ])
             .unwrap()
-        )
+        );
     }
 }