+++ cedar-policy/src/integration_testing.rs
@@ -143,7 +143,7 @@ pub trait CustomCedarImpl {
         q: &cedar_policy_core::ast::Request,
         p: &cedar_policy_core::ast::PolicySet,
         e: &cedar_policy_core::entities::Entities,
-    ) -> cedar_policy_core::authorizer::Response;
+    ) -> Response;
 
     /// Custom validator entry point.
     ///
@@ -284,9 +284,7 @@ pub fn perform_integration_test_from_json_custom(
             });
         let request = Request::new(principal, action, resource, context);
         let response = if let Some(custom_impl) = custom_impl_opt {
-            custom_impl
-                .is_authorized(&request.0, &policies.ast, &entities.0)
-                .into()
+            custom_impl.is_authorized(&request.0, &policies.ast, &entities.0)
         } else {
             Authorizer::new().is_authorized(&request, &policies, &entities)
         };
@@ -323,7 +321,9 @@ pub fn perform_integration_test_from_json_custom(
             let mut found_matching_non_existent_fn_fuzzing = false;
             for e in expected_response.diagnostics().errors() {
                 let EvaluationError::StringMessage(msg) = e;
-                if msg.contains("while evaluating policy policy0, encountered the following error: function does not exist:") {
+                if msg.contains(
+                    "error occurred while evaluating policy `policy0`: function does not exist:",
+                ) {
                     let fuzzing_fn_name = Some(msg.split_whitespace().last().unwrap().to_string());
                     if parsing_fn_name == fuzzing_fn_name {
                         found_matching_non_existent_fn_fuzzing = true;
@@ -355,7 +355,8 @@ pub fn perform_integration_test_from_json_custom(
         let ests = policies
             .policies()
             .map(|p| p.to_json().expect("should convert to JSON successfully"));
-        let _ = PolicySet::from_policies(ests.enumerate().map(|(i, est)| {
+
+        PolicySet::from_policies(ests.enumerate().map(|(i, est)| {
             let id = PolicyId::from_str(&format!("policy{i}")).expect("id should be valid");
             Policy::from_json(Some(id), est.clone()).unwrap_or_else(|e| {
                 panic!("in test {}, failed to build policy from JSON successfully: {e}\n\ntext policy was:\n{}\n\nJSON policy was: {}\n",